+++ apps/sim/app/workspace/[workspaceId]/files/components/file-viewer/rich-markdown-editor/code-block.tsx
@@ -8,12 +8,11 @@ import {
   DropdownMenuTrigger,
   useCopyToClipboard,
 } from '@sim/emcn'
-import type { JSONContent } from '@tiptap/core'
-import { CodeBlock } from '@tiptap/extension-code-block'
 import type { ReactNodeViewProps } from '@tiptap/react'
 import { NodeViewContent, NodeViewWrapper, ReactNodeViewRenderer } from '@tiptap/react'
 import { Check, ChevronDown, Code, Copy, Eye, WrapText } from 'lucide-react'
 import { looksLikeMermaid, MermaidDiagram } from '../mermaid-diagram'
+import { MarkdownCodeBlock } from './code-block-schema'
 import { detectLanguage } from './detect-language'
 import { useEditorEditable } from './use-editor-editable'
 
@@ -227,30 +226,6 @@ function CodeBlockView({ node, updateAttributes, editor, getPos }: ReactNodeView
     </NodeViewWrapper>
   )
 }
-
-function codeBlockText(node: JSONContent): string {
-  return (node.content ?? []).map((child) => child.text ?? '').join('')
-}
-
-/** Fence sized to one backtick longer than the longest run inside the code (CommonMark rule). */
-function fenceFor(text: string): string {
-  const longestRun = Math.max(0, ...[...text.matchAll(/`+/g)].map((match) => match[0].length))
-  return '`'.repeat(Math.max(3, longestRun + 1))
-}
-
-/**
- * Code block whose markdown serializer sizes the fence to the interior backtick runs, so a code
- * block that itself contains a ``` line round-trips instead of shattering. Shared by the test
- * (plain) and live ({@link CodeBlockWithLanguage}) paths.
- */
-export const MarkdownCodeBlock = CodeBlock.extend({
-  renderMarkdown: (node: JSONContent) => {
-    const language = typeof node.attrs?.language === 'string' ? node.attrs.language : ''
-    const text = codeBlockText(node)
-    const fence = fenceFor(text)
-    return `${fence}${language}\n${text}\n${fence}`
-  },
-})
 
 /**
  * Code block with hover-revealed controls (language picker, line-wrap toggle, copy). The
